[TRL-430] refactor: 인증 관련 Exception 재정의
