+++ src/main/java/com/cosain/trilo/common/exception/auth/AccessTokenNotExistException.java
@@ -3,7 +3,7 @@ package com.cosain.trilo.common.exception.auth;
 import com.cosain.trilo.common.exception.CustomException;
 import org.springframework.http.HttpStatus;
 
-public class TokenNotExistException extends CustomException {
+public class AccessTokenNotExistException extends CustomException {
 
     private static final String ERROR_CODE = "auth-0001";
     private static final HttpStatus HTTP_STATUS = HttpStatus.UNAUTHORIZED;
