+++ src/main/java/com/cosain/trilo/common/exception/auth/RefreshTokenNotExistException.java
@@ -1,10 +1,11 @@
-package com.cosain.trilo.common.exception;
+package com.cosain.trilo.common.exception.auth;
 
+import com.cosain.trilo.common.exception.CustomException;
 import org.springframework.http.HttpStatus;
 
-public class NotExistRefreshTokenException extends CustomException {
+public class RefreshTokenNotExistException extends CustomException {
 
-    private static final String ERROR_CODE = "auth-0007";
+    private static final String ERROR_CODE = "auth-0004";
     private static final HttpStatus HTTP_STATUS = HttpStatus.UNAUTHORIZED;
 
     @Override
