+++ src/main/java/com/cosain/trilo/common/exception/auth/RefreshTokenNotValidException.java
@@ -3,9 +3,9 @@ package com.cosain.trilo.common.exception.auth;
 import com.cosain.trilo.common.exception.CustomException;
 import org.springframework.http.HttpStatus;
 
-public class TokenInvalidFormatException extends CustomException {
+public class RefreshTokenNotValidException extends CustomException {
 
-    private static final String ERROR_CODE = "auth-0003";
+    private static final String ERROR_CODE = "auth-0005";
     private static final HttpStatus HTTP_STATUS = HttpStatus.UNAUTHORIZED;
 
     @Override
